requirements.txt
asgiref
Django
sqlparse
tzdata
django-filter